See description below!
I’ve added enemy spawners to the game. When the game starts, a random number of enemies will spawn, setting the maximum enemy count for that session. This maximum will only reset when you quit and restart the game, causing a new random number of enemies to appear.

For example, if 10 enemies spawn initially, that’s the max number you’ll have for that session. If you kill 5 of them, only those 5 will respawn, keeping the total count at 10. The game will continue respawning only the number of enemies you’ve eliminated, up to the maximum set at the start.